+++ src/main/java/com/onz/bars/bar/component/apiCaller/ApiCaller.java
@@ -1,7 +1,7 @@
 package com.onz.bars.bar.component.apiCaller;
 
 
-import com.onz.bars.bar.domain.apiResponse.ApiResponse;
+import com.onz.bars.bar.model.response.ApiResponse;
 
 public interface ApiCaller {
