--- src/main/java/com/onz/bars/bar/domain/apiResponse/ApiResponse.java
@@ -1,4 +1,0 @@
-package com.onz.bars.bar.domain.apiResponse;
-
-public abstract class ApiResponse {
-}
