+++ src/main/java/com/onz/bars/bar/model/response/ApiResponse.java
@@ -1,0 +1,4 @@
+package com.onz.bars.bar.model.response;
+
+public abstract class ApiResponse {
+}
