+++ src/main/java/com/onz/bars/bar/model/response/googlePlace/placeDetailApi/GooglePlaceDetailApiResponse.java
@@ -1,6 +1,6 @@
-package com.onz.bars.bar.domain.apiResponse.googlePlace.placeDetailApi;
+package com.onz.bars.bar.model.response.googlePlace.placeDetailApi;
 
-import com.onz.bars.bar.domain.apiResponse.ApiResponse;
+import com.onz.bars.bar.model.response.ApiResponse;
 
 import java.util.List;
