+++ src/main/java/com/onz/bars/bar/model/response/googlePlace/textSearchApi/Place.java
@@ -1,4 +1,4 @@
-package com.onz.bars.bar.domain.apiResponse.googlePlace.textSearchApi;
+package com.onz.bars.bar.model.response.googlePlace.textSearchApi;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
