+++ src/main/java/com/onz/bars/bar/service/KakaoLocalApiService.java
@@ -2,8 +2,8 @@ package com.onz.bars.bar.service;
 
 import com.onz.bars.bar.component.apiCaller.KakaoLocalApiCaller;
 import com.onz.bars.bar.domain.Bar;
-import com.onz.bars.bar.domain.apiResponse.kakaoLocal.KakaoMapApiResponse;
-import com.onz.bars.bar.domain.apiResponse.kakaoLocal.Place;
+import com.onz.bars.bar.model.response.kakaoLocal.KakaoMapApiResponse;
+import com.onz.bars.bar.model.response.kakaoLocal.Place;
 import com.onz.bars.bar.domain.menu.Menu;
 import com.onz.bars.bar.domain.openHour.OpenHour;
 import com.onz.bars.bar.repository.BarRepository;
@@ -12,7 +12,7 @@ import com.onz.bars.bar.repository.OpenHourReposistory;
 import com.onz.bars.bar.component.scraper.KakaoMapMenuMapper;
 import com.onz.bars.bar.component.scraper.KakaoMapOpenHourMapper;
 import com.onz.bars.bar.component.scraper.KakaoMapScraper;
-import com.onz.bars.bar.domain.apiResponse.scrapeResponse.KakaoMapScrapeResponse;
+import com.onz.bars.bar.model.response.scrapeResponse.KakaoMapScrapeResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
